+++ gulp/res/js/importexport.js
@@ -4,7 +4,7 @@ window.addEventListener('settingsReady', () => {
 	const settingNames = ['volume','loop','imageloadingbars','live','scroll','localtime','relative','24hour','notifications','hiddenimages', 'threadwatcher'
 		,'notification-yous-only','yous-setting','filters1','name','theme','codetheme','customcss','disableboardcss','hiderecursive', 'watchlist'
 		,'heightlimit','crispimages','hidethumbnails','noncolorids','alwaysshowspoilers','hidedeletedpostcontent','hidepoststubs','smoothscrolling'
-		,'tegakiheight-setting','tegakiwidth-setting'];
+		,'tegakiheight-setting','tegakiwidth-setting', 'hoverexpandsmedia'];
 
 	const importExportText = document.getElementById('import-export-setting');